@@ -2,12 +2,13 @@
 
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 	<GenerateAssemblyConfigurationAttribute>false</GenerateAssemblyConfigurationAttribute>
 	<GenerateAssemblyCompanyAttribute>false</GenerateAssemblyCompanyAttribute>
 	<GenerateAssemblyProductAttribute>false</GenerateAssemblyProductAttribute>
 	<Description>Opux is a bot for Discord That manages EVE Online things</Description>
 	<NeutralLanguage>en-GB</NeutralLanguage>
+	<StartupObject>Opux.Program</StartupObject>
   </PropertyGroup>
 
   <ItemGroup>
@@ -21,7 +22,7 @@
   <ItemGroup>
     <PackageReference Include="Bugsnag" Version="3.1.0" />
     <PackageReference Include="ByteSize" Version="2.1.1" />
-    <PackageReference Include="Discord.Net" Version="2.1.1" />
+    <PackageReference Include="Discord.Net" Version="3.9.0" />
     <PackageReference Include="ESIClient.Dotcore" Version="1.0.13" />
     <PackageReference Include="log4net" Version="2.0.14" />
     <PackageReference Include="MatriX.NetCore" Version="2.0.0.22-beta4" />
